ユーザー編集処理実装(追加はまだよ）

+++ WebApplication1/WebApplication1/Controllers/UsersController.cs
@@ -31,7 +31,7 @@ namespace WebApplication1.Controllers
 
         // GET: api/Users/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<User>> GetUser(long id)
+        public async Task<ActionResult<User>> GetUser(int id)
         {
             var user = await _context.Users.FindAsync(id);
 
@@ -45,7 +45,7 @@ namespace WebApplication1.Controllers
 
         // PUT: api/Users/5
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutUser(long id, User user)
+        public async Task<IActionResult> PutUser(int id, User user)
         {
             if (id != user.Id)
             {
